Made module to get expressions
